Update lab1.cpp
Исправил тип данных с size_t на  int

+++ Lab1CPP/lab1.cpp
@@ -6,7 +6,7 @@
 
 void task1(Array* arr)
 {
-    size_t result = 0;
+    int result = 0;
 
     for (size_t i = 0; i < array_size(arr); i++)
     {
